move basic mempool operations to blockchain struct, remove duplicate functions, remove mempool_remove_transactions as it was only used once and the behaviour was incorrect
