+++ crates/l2/proposer/l1_watcher.rs
@@ -4,7 +4,7 @@ use crate::{
 };
 use bytes::Bytes;
 use ethereum_types::{Address, BigEndianHash, H256, U256};
-use ethrex_blockchain::{constants::TX_GAS_COST, mempool, Blockchain};
+use ethrex_blockchain::{constants::TX_GAS_COST, Blockchain};
 use ethrex_common::types::{Signable, Transaction};
 use ethrex_rpc::clients::eth::{errors::EthClientError, eth_sender::Overrides, EthClient};
 use ethrex_rpc::types::receipt::RpcLog;
@@ -267,10 +267,9 @@ impl L1Watcher {
                 .await?;
             mint_transaction.sign_inplace(&self.l2_proposer_pk);
 
-            match mempool::add_transaction(
-                Transaction::PrivilegedL2Transaction(mint_transaction),
-                blockchain,
-            ) {
+            match blockchain
+                .add_transaction_to_pool(Transaction::PrivilegedL2Transaction(mint_transaction))
+            {
                 Ok(hash) => {
                     info!("Mint transaction added to mempool {hash:#x}",);
                     deposit_txs.push(hash);
